+++ src/main/java/io/nut/base/stats/BigSimpleMovingAverage.java
@@ -1,7 +1,7 @@
 /*
- *  SimpleMovingAverage.java
+ *  BigSimpleMovingAverage.java
  *
- *  Copyright (c) 2024-2025 francitoshi@gmail.com
+ *  Copyright (c) 2024-2026 francitoshi@gmail.com
  *
  *  This program is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -31,7 +31,7 @@ import java.util.concurrent.ArrayBlockingQueue;
  */
 public class BigSimpleMovingAverage extends BigMovingAverage 
 {
-    private int count;
+    private long count;
     private final int period;
     private final BigDecimal p;
     private final int scale;
