+++ src/main/java/io/nut/base/stats/BigWeightedMovingAverage.java
@@ -1,7 +1,7 @@
 /*
- *  WeightedMovingAverage.java
+ *  BigWeightedMovingAverage.java
  *
- *  Copyright (c) 2024-2025 francitoshi@gmail.com
+ *  Copyright (c) 2024-2026 francitoshi@gmail.com
  *
  *  This program is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -28,7 +28,7 @@ import java.util.List;
 public class BigWeightedMovingAverage extends BigMovingAverage
 {
 
-    private int count;
+    private long count;
     private final int period;
     private final int scale;
     private final RoundingMode roundingMode;
