+++ src/main/java/io/nut/base/stats/SimpleMovingAverage.java
@@ -1,7 +1,7 @@
 /*
  *  SimpleMovingAverage.java
  *
- *  Copyright (c) 2024-2025 francitoshi@gmail.com
+ *  Copyright (c) 2024-2026 francitoshi@gmail.com
  *
  *  This program is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -25,7 +25,7 @@ import java.util.concurrent.ArrayBlockingQueue;
 
 public class SimpleMovingAverage extends MovingAverage 
 {
-    private int count;
+    private long count;
     private final int period;
     
     private double sum = 0.0;
