+++ src/main/java/io/nut/base/stats/WeightedMovingAverage.java
@@ -1,7 +1,7 @@
 /*
  *  WeightedMovingAverage.java
  *
- *  Copyright (c) 2024-2025 francitoshi@gmail.com
+ *  Copyright (c) 2024-2026 francitoshi@gmail.com
  *
  *  This program is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -26,7 +26,7 @@ import java.util.List;
 public class WeightedMovingAverage extends MovingAverage
 {
 
-    private int count;
+    private long count;
     private final int period;
     private final List<Double> values;
     private double sum;
